requirements.txt
langchain
langchain-anthropic
langchain-google-genai
langgraph
pillow
pydantic
python-dotenv